Add support for resetting each work directory
Previously the work directory was always called "work" and the stage
skip files were always in "stages", but that's no longer the case,
the skip files are in each work directory to keep them separate and
the work directories are named by arch+distro.

By default we'll reset the most commonly built image, and show a help
message if the person building doesn't know to add arguments for the
arch and distro.

+++ reset.sh
@@ -1,7 +1,27 @@
 #!/bin/bash -e
 
-pushd stages
-find . -type f -name 'SKIP' -delete
-find . -type f -name 'SKIP_STEP*' -delete
-find . -type f -name 'SKIP_CH_STEP*' -delete
-popd
+IMAGE_ARCH=$1
+DISTRO=$2
+
+echo "Reset Open.HD Builder"
+echo "------------------------------------------------------"
+echo ""
+
+
+if [[ "${IMAGE_ARCH}" == "" || ${DISTRO} == "" ]]; then
+    IMAGE_ARCH="pi"
+    DISTRO="stretch"
+
+    echo "Usage: ./reset.sh pi [stretch | buster]"
+    echo "------------------------------------------------------"
+    echo ""
+fi
+
+if [ -d "work-${IMAGE_ARCH}-${DISTRO}" ]; then
+    pushd "work-${IMAGE_ARCH}-${DISTRO}"
+        find . -type f -name 'SKIP' -delete
+        find . -type f -name 'SKIP_STEP*' -delete
+        find . -type f -name 'SKIP_CH_STEP*' -delete
+        find . -type f -name 'IMAGE.img' -delete
+    popd
+fi
